calc. triangle: fix cases where 2nd sol. is valid and 1st not (fix #760)
In some cases, the second computed solution is valid and the first one
not. In this case, we need to swap solutions and declare only the first
one is valid.
Note that the function to check if the first solution is valid was not
good enough since it ommited checking for angles < 0.0. This has been
addressed.

While there, refactor things a bit.