jobs: improve job import process
The job activity now displays a list of all tpst files from the
'TopoSuite' directory. A context menu is used to import a job from the
selected tpst. The other action buttons ('export current job', 'clear
current job') have been placed to a new option menu.

Note that this is likely to change in the future as part of a job
management overhaul for the next major version of TopoSuite.