points: use cache dir for share action (close #841)
We previously created a temporary file on external storage when sharing
a file of points was needed. This is bad practice. Moreover, there is no
need to get write permission for sharing a file. Instead, we can simply
use internal storage and give temporary read permission when sharing via
intent. Hence, this is also safer.

As a consequence, I also decided to remove the default "tmp dir" set in
App. This is not needed anymore. By doing so, we avoid a lot of
problems, notably on Android 6.0+ where it is required to ask for
permission to write to external storage.