make 'Calculation.recordToHistory()' private
This method has no reason to be public. We actually need to serialize
calculations via bundle to pass them to other activities. Using the
history for this purpose is PLAIN WRONG, dammit!
Hence, adapt orthogonal survey and axis implantation calculation to use
bundle to serialize calculation to pass it to the result activity. Of
course, this requires the calculation to be plainly serializable.